Remove duplicated read_yaml_file function
