+++ unittest/force-styles/angle_style.cpp
@@ -193,18 +193,6 @@ void data_lammps(LAMMPS_NS::LAMMPS *lmp, const TestConfig &cfg)
 }
 
 
-// read/parse yaml file
-
-bool read_yaml_file(const char *infile, TestConfig &config)
-{
-    auto reader = TestConfigReader(config);
-    if (reader.parse_file(infile))
-        return false;
-    
-    config.basename = reader.get_basename();
-    return true;
-}
-
 // re-generate yaml file with current settings.
 
 void generate_yaml_file(const char *outfile, const TestConfig &config)
